Fixed the green thresholding
I ran my video script with a video of green roombas and discovered that my green thresholding was not amazing. It still isn't amazing, but it is now ignoring things more things that it was before. I also fixed something stupid I was doing with the red thresholding so that red swtiches show up as full rectangles instead of vague outlines.

Also added a color spectrum analyzer so that we can test what actually happens when we use inRange.