This is a basic GHT test that should work.
I tried using two different templates - one called blackOutline.png, which is just a basic outline of the roomba switch in black and white that I thought would work as a template, and another one called altTemplate.png, which is just a picture of the roomba switch with a little bit of the roomba surrounding it. Neither of them have yielded any results with testGeneralHough.png.